Fix texture clipping (!!!!!)

+++ src/graphics/Rasterizer.java
@@ -10,7 +10,6 @@ import java.util.Comparator;
 import java.util.stream.IntStream;
 
 public class Rasterizer {
-    //sep class because it's also gonna handle depth buffer, perspective stuff
     public static void drawTexTriangle(Graphics g, Triangle tri, int width, int height) {
         BufferedImage bi;
         try {
@@ -34,7 +33,6 @@ public class Rasterizer {
         int y1 = (int)ptsClone[0].y; int y2 = (int)ptsClone[1].y; int y3 = (int)ptsClone[2].y;
         int u1 = (int)texPtsClone[0].x; int u2 = (int)texPtsClone[1].x; int u3 = (int)texPtsClone[2].x;
         int v1 = (int)texPtsClone[0].y; int v2 = (int)texPtsClone[1].y; int v3 = (int)texPtsClone[2].y;
-        //System.out.println(Arrays.toString(texPtsClone));
 
         int dx1 = x2 - x1; int dy1 = y2 - y1;
         double du1 = u2 - u1; double dv1 = v2 - v1;
@@ -54,23 +52,21 @@ public class Rasterizer {
             dU2Step = du2 /Math.abs(dy2);
             dV2Step = dv2 /Math.abs(dy2);
         }
-        //System.out.println(dU1Step);
 
         for (int z = 0; z < 2; z++) {
             if (dy1 != 0) {
                 for (int i = (z==0 ? y1 : y2); i <= (z==0 ? y2: y3); i++) {
-                    //if (i >= height) break;
 
                     int startX = z==0 ? (x1 + (int) ((i - y1) * dStartXStep)) :
                             (x2 + (int) ((i - y2) * dStartXStep));
                     int endX = x1 + (int) ((i - y1) * dEndXStep);
 
-                    double startU = z==0 ? (u1 + ((i-y1)*dU1Step)) :
-                            (u2 + ((i - y2) * dU1Step));
-                    double startV = z==0 ? (v1 + ((i-y1)*dV1Step)) :
-                            (v2 + ((i - y2) * dV1Step));
-                    double endU = u1 + ((i - y1) * dU2Step);
-                    double endV = v1 + ((i - y1) * dV2Step);
+                    double startU = z==0 ? (u1 + ((double)(i-y1)*dU1Step)) :
+                            (u2 + ((double)(i - y2) * dU1Step));
+                    double startV = z==0 ? (v1 + ((double)(i-y1)*dV1Step)) :
+                            (v2 + ((double)(i - y2) * dV1Step));
+                    double endU = u1 + (double)(i - y1) * dU2Step;
+                    double endV = v1 + (double)(i - y1) * dV2Step;
 
                     //need to sort along x
                     if (startX > endX) {
@@ -88,13 +84,8 @@ public class Rasterizer {
                     double u, v;
                     double tStep = 1.0 / (endX - startX);
                     double t = 0;
-                    //System.out.println("start and end U " + startU + " " + endU);
-                    //System.out.println("start and end V " + startV + " " + endV);
 
                     for (int j = startX; j < endX; j++) {
-                        //if (j >= width) break;
-                        //linear interpolation, weighted average version
-                        //System.out.println(t);
                         u = (1 - t) * startU + t * endU;
                         v = (1 - t) * startV + t * endV;
 
@@ -104,12 +95,8 @@ public class Rasterizer {
                         //bitwise rgb color value
                         Color c = new Color((color & 0xff0000) >> 16, (color & 0xff00) >> 8, color & 0xff);
 
-                        //System.out.println(u + " " + v + " " + c);
-
                         g.setColor(c);
                         g.drawRect(width-j, height-i, 1, 1);
-
-                        //System.out.println(j + " " + i);
 
                         t += tStep;
                     }
